+++ src/vue/App.vue
@@ -101,6 +101,7 @@ export default {
       vuexTypes.isAccountBlocked,
       vuexTypes.kycRequestBlockReason,
       vuexTypes.account,
+      vuexTypes.isAccountCorporate,
     ]),
     isNavigationRendered () {
       return this.$route.matched.some(m => m.meta.isNavigationRendered)
@@ -144,7 +145,7 @@ export default {
       startIdle: vuexTypes.START_IDLE,
       logoutSession: vuexTypes.LOGOUT_SESSION,
       restoreSession: vuexTypes.RESTORE_SESSION,
-      loadBusinessStatsQuoteAsset: vuexTypes.LOAD_BUSINESS_STATS_QUOTE_ASSET,
+      loadBusiness: vuexTypes.LOAD_BUSINESS,
     }),
     ...mapMutations({
       popState: vuexTypes.POP_STATE,
@@ -165,8 +166,8 @@ export default {
       walletsManager.useApi(api)
       factorsManager.useApi(api)
       documentsManager.useApi(api)
+      if (this.isAccountCorporate) this.loadBusiness(this.walletAccountId)
       await this.loadAssets()
-      await this.loadBusinessStatsQuoteAsset()
     },
     detectIncompatibleBrowser () {
       this.isNotSupportedBrowser = !isCompatibleBrowser()