+++ src/vue/fields/SelectField.vue
@@ -233,10 +233,6 @@ export default {
   overflow-y: auto;
   padding: 0.8rem 0;
 
-  @include respond-to-custom($sidebar-hide-bp) {
-    top: auto;
-    bottom: 0;
-  }
   @include box-shadow;
 }