+++ src/vue/forms/LoginForm.vue
@@ -102,7 +102,6 @@ export default {
     ...mapActions({
       logInAccount: vuexTypes.LOG_IN,
       loadAssets: vuexTypes.LOAD_ASSETS,
-      loadBusinessStatsQuoteAsset: vuexTypes.LOAD_BUSINESS_STATS_QUOTE_ASSET,
     }),
     async submit () {
       if (!this.isFormValid()) return
@@ -115,7 +114,6 @@ export default {
           password: this.form.password,
         })
         await this.loadAssets()
-        await this.loadBusinessStatsQuoteAsset()
         await this.$router.push({ name: 'app' })
       } catch (e) {
         this.processAuthError(e)