+++ CHANGELOG.md
@@ -10,19 +10,46 @@ for further information about branching and tagging conventions.
 #### Added
 - Trim for input field 
 - Redeem default amount
+- Current company page
+- New component 'Card' for displaying all card
+- Skeleton loader for new component 'Card'
+- Global class for markdown `app__markdown`
+
+#### Removed
+- `card-viewer`
+- `asset-skeleton-loade`
+- `current-business-logo`
+- `current-business-indicator/index.vue`
+- `BusinessCardSkeleton`
+- `BusinessViewer`
+- `PromoCodeCardSkeleton`
 
 #### Changed
 - Moved logo and favicon to branding folder
 - Display full price
 - Max percent discount
 - Validation on promo code form
+- Get data from store on Business pages
 
 #### Fixed
 - A bug with trim in input fields with type 'password'
+- A bug with typo in multi select
+- A bug with color scheme
+- A bug when don't show error message if response failed
 
 ### "Under the hood" changes
 #### Added
 - `APP_NAME` in config
+
+## [1.3.0-rc.0] - 2019-10-31
+### "Under the hood" changes
+#### Changed
+- Now using @tokend/js-sdk@1.11.1
+
+## [1.2.0] - 2019-10-31
+#### Fixed
+- A bug with skeleton loader on Shop page
+- A bug when don't show error message if response failed
 
 ## [1.2.0-rc.0] - 2019-10-21
 #### Added
@@ -359,7 +386,9 @@ for further information about branching and tagging conventions.
 #### Removed
 - `RECOVERY_MODE` config key due to unused anymore
 
-[Unreleased]: https://github.com/tokend/web-client/compare/1.2.0-rc.0...HEAD
+[Unreleased]: https://github.com/tokend/web-client/compare/1.3.0-rc.0...HEAD
+[1.3.0-rc.0]: https://github.com/tokend/web-client/compare/1.2.0...1.3.0-rc.0
+[1.2.0]: https://github.com/tokend/web-client/compare/1.2.0-rc.0...1.2.0
 [1.2.0-rc.0]: https://github.com/tokend/web-client/compare/1.1.0...1.2.0-rc.0
 [1.1.0]: https://github.com/tokend/web-client/compare/1.1.0-rc.3...1.1.0
 [1.1.0-rc.3]: https://github.com/tokend/web-client/compare/1.1.0-rc.2...1.1.0-rc.3